refactor(clawd): Extract subcommands into dedicated files
Moves the `Run` and `Doctor` command implementations from the main `Clawd.swift`
file into their own respective `RunCommand.swift` and `DoctorCommand.swift`
files within a new `Subcommands` directory.

This refactoring improves modularity and organization by reducing the size
and complexity of the primary application entry point.
